+++ habitat/utils/pickle5_multiprocessing.py
@@ -8,6 +8,7 @@ import io
 import sys
 from multiprocessing.connection import Connection
 from multiprocessing.reduction import ForkingPickler as _ForkingPickler
+from typing import Callable
 
 from habitat.core.logging import logger
 
@@ -75,3 +76,25 @@ class ConnectionWrapper:
                 type(self).__name__, "conn"
             )
         )
+
+
+class CloudpickleWrapper:
+    """Wrapper that uses cloudpickle to pickle and unpickle the result."""
+
+    def __init__(self, fn: Callable):
+        """Cloudpickle wrapper for a function."""
+        self.fn = fn
+
+    def __getstate__(self):
+        """Get the state using `cloudpickle.dumps(self.fn)`."""
+        import cloudpickle
+
+        return cloudpickle.dumps(self.fn)
+
+    def __setstate__(self, ob):
+        """Sets the state with obs."""
+        self.fn = pickle.loads(ob)
+
+    def __call__(self, *args, **kwargs):
+        """Calls the function `self.fn` with no arguments."""
+        return self.fn(*args, **kwargs)
